Removed conflicting overloaded methods from RestTemplate
RestTemplate consisted of methods of the following form (getFor..., postFor..., etc.)
    def m(...)
and
    def m(..., uriVariables: Any*)

This made it impossible to call methods without passing any uriVariables, since the compiler cannot determine which of the overloaded method is intended. In any case, there is no need for the non-varargs methods; users can just call the varargs version passing no arguments for the uriVariables parameter.